backend/requirements.txt
Flask
Flask-Cors
spacy
mysql-connector-python
openai
python-dotenv